checkpoint cleanup tests
