+++ bin/user/tests/test_sql_statements.py
@@ -155,6 +155,35 @@ def setup_config(calculated_field, input_field, algorithm, aqi_type):
     }
     return config_dict
 
+class Test01(unittest.TestCase):
+    def test_01(self):
+        mock_logger = mock.Mock(spec=user.aqitype.Logger)
+
+        config = configobj.ConfigObj(io.StringIO(MANAGER_DICT))
+        db_binder = weewx.manager.DBBinder(config)
+        db_manager = db_binder.get_manager('aqi_binding')
+
+        SUT = user.aqitype.NOWCAST(mock_logger, random.randint(1, 100), random_string(), 'pm2_5')
+
+        # The timestamp in the template that I am using is 1740168300 (3:05 PM Eastern on 2/21/2025)
+        # The code finds the hour that this timestamp belongs to, it is 1740168000 (3:00 PM Eastern on 2/21/2025)
+        # This is the stop value that is passed into _get_concentration_data
+        # This is used as the stop value in the SQL queries
+        ret_value = SUT._get_concentration_data(db_manager, 1740168000)
+
+        expected_pm_values = (1.3857366771159874, 1.4406226053639848, 1.5153544061302677,
+                              1.7157567049808427, 1.4154310344827585, 1.5376532567049808,
+                              1.3357950191570878, 1.5952873563218388, 1.3942241379310343,
+                              1.4037739463601533, 1.1982950191570885, 1.2343007662835248)
+        expected_timestamps = (1740171300, 1740168000, 1740164400, 1740160800, 1740157200, 1740153600,
+                               1740150000, 1740146400, 1740142800, 1740139200, 1740135600, 1740132000)
+
+        self.assertEqual(ret_value,
+                         (12, 1.1982950191570885, 1.7157567049808427, expected_timestamps, expected_pm_values)
+                        )
+
+        db_binder.close()
+
 class TestEPAAQICalculate(unittest.TestCase):
     def test_get_series_data(self):
         mock_logger = mock.Mock(spec=user.aqitype.Logger)
